refactor: ♻️ rename repository definition
