+++ libs/tasks/domain/src/index.ts
@@ -1,6 +1,10 @@
-export * from './lib/entities/task.entity';
-export * from './lib/entities/user-task.entity';
+// tasks
+export * from './lib/tasks-repository.interface';
 
+export * from './lib/entities/task.entity';
 export * from './lib/value-objects/task-status.enum';
 
-export * from './lib/tasks.repository';
+// user tasks
+export * from './lib/user-tasks-repository.interface';
+
+export * from './lib/entities/user-task.entity';
